fix: do not use workspace name as project name by default
The Snyk CLI implicitly detects the project name from the manifest file and the folder of the project. This can be overridden by the `--project-name` argument.
Removing the default behaviour where the workspace name is used as the project name to allow the implicit project name detection as the default behaviour.