chore: update pom dependency, base image to jenkins 2.442
Mitigate critical severity vuln
https://app.snyk.io/vuln/SNYK-JAVA-ORGJENKINSCIMAIN-6190606
by updating the pom dependency and base image to Jenkins 2.442.

In this new base image, system-wide pip installs aren't allowed so the Debian
packages are updated to satisfy the virtualenv requirement.

Installing Debian packages with --no-install-recommends because the default was
pulling in x11 and a bunch of extra junk a Jenkins server shouldn't need.

Drive-by: fix docker command-line flag in run script, noticed this when testing
the image locally.